Alter how usages are totalized in pge-urjanet-generation - we typically have to totalize the usage ourselves with the xml datasource. (#285)
Treat a usage with a RateComponent as ["total"] differently that a RateComponent as [total, winter] / [total, summer], etc.  For the latter case, for svce, the customer was on GreenPrime, and usages were coming in under [total, winter] and [total, summer] but the only the first usage was counted.

Fixed an existing bug in a test where this applied.