Partial Urjanet Scrapers - Scrape Date Range [GRID-3443] (#290)
* For partial billing scrapers, restrict the dates scraped to bills within the requested date range.

For partial scrapers, especially where we're scraping historical data, it is useful to retrieve the same amount of data from both datasources.  Because Urjanet always returns everything we have by default, if there is a big mismatch between this and the amount of data requested from the other data source, we could lose a lot of data from the non-urjanet source.

Doing this at the very end so all urjanet bills, including urjanet bills, get incorporated into the billing history, and then we just chop off what we don't want.